Significant progress towards basic game.
Movement and collisions work. Since robots are just map items (not entities tracked otherwise) the mechanism for identifying and moving bots is a little funky. If I just loop through the map moving robots when I encounter them I will move some bots multiple times in one round as they are re-encountered in the map in their new position. So, I put all bots in a list and only move each list itemonce. There is a chance that a bot in the list will be destroyed before its time, so we have to check for that.

I also implemented teleport (t) and diagonal movement (q,e,z,c).

Teleport spamming seems to result in collisions (not fragging robots).

FIXME / CHECKME: what happens if you teleport onto a staircase?

Is scoring fair / interesting?

FIXME: need sensors and other guidance variables so that it can be played via little python