Return early from handle_keys upon descent/mapgen
There was a bug where the new map was generated when a player descended
the stairs where it would appear that a robot was spawning on top of the
player (on the new level).

Actually, __create_map placed the player last to specifically avoid
telefragging. However, in handle_keys we did not return early after a
new map generation, which meant that after the player descended the
stairs, a new cadre of robots was randomly generated, and if one spawned
*next* to the player it had the opportunity to move immediately (a "free
attack" in RPG parlance).

I fixed this bug by returning immediately after calling __create_map in
handle_keys. This means that the player gets the first move on each new
level, but that was the original intent, anyway.